refactor meta resolver file name and classes
