+++ projects/core/src/cms/page/index.ts
@@ -1,2 +1,2 @@
 export * from './page-meta.resolver';
-export * from './content-page-title.resolver';
+export * from './content-page-meta.resolver';
