+++ projects/core/src/cms/page/page.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { PageMetaResolver } from './page-meta.resolver';
-import { ContentPageTitleResolver } from './content-page-title.resolver';
+import { ContentPageMetaResolver } from './content-page-meta.resolver';
 
 @NgModule({
   providers: [
     {
       provide: PageMetaResolver,
-      useExisting: ContentPageTitleResolver,
+      useExisting: ContentPageMetaResolver,
       multi: true
     }
   ]
