+++ projects/core/src/product/services/index.ts
@@ -1,3 +1,3 @@
-export * from './category-page-title.resolver';
-export * from './product-page-title.resolver';
-export * from './search-page-title.resolver';
+export * from './category-page-meta.resolver';
+export * from './product-page-meta.resolver';
+export * from './search-page-meta.resolver';
